Extract an interface for SoundSet as prep for test-dummy SoundSets.
An interface for something with only one sane implementation (SampledSoundSet)
might seem like overkill, but this allows us to write tests without adding
a dependency on a mocking framework, which would be worse overkill.
